cambio de tamaño de los dias de la semana de "long" a "short"

+++ src/funcionalities/MonthSelector.ts
@@ -47,7 +47,7 @@ export function initializeCalendar() {
 
     for (let i = 0; i < 7; i++) {
       const date = new Date(currentYear, monthNumber, i + 1);
-      const weekday = new Intl.DateTimeFormat(navigator.language, { weekday: "long" }).format(date);
+      const weekday = new Intl.DateTimeFormat(navigator.language, { weekday: "short" }).format(date);
       weekTotalDays.push(weekday);
     }
